chore: creating fcc-store-sql module
