+++ extensions/store/fcc-store-sql/src/main/java/de/truzzt/edc/extension/catalog/cache/sql/schema/postgres/PostgresDialectStatements.java
@@ -12,9 +12,9 @@
  *
  */
 
-package org.eclipse.edc.catalog.store.sql.schema.postgres;
+package de.truzzt.edc.extension.catalog.cache.sql.schema.postgres;
 
-import org.eclipse.edc.catalog.store.sql.schema.BaseSqlDialectStatements;
+import de.truzzt.edc.extension.catalog.cache.sql.schema.BaseSqlDialectStatements;
 import org.eclipse.edc.spi.query.QuerySpec;
 import org.eclipse.edc.sql.dialect.PostgresDialect;
 import org.eclipse.edc.sql.translation.SqlQueryStatement;
