+++ extensions/store/fcc-store-sql/src/main/java/de/truzzt/edc/extension/catalog/cache/sql/schema/postgres/SelectorExpressionMapping.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.edc.catalog.store.sql.schema.postgres;
+package de.truzzt.edc.extension.catalog.cache.sql.schema.postgres;
 
 import org.eclipse.edc.spi.asset.AssetSelectorExpression;
 import org.eclipse.edc.sql.translation.TranslationMapping;
