+++ launchers/build.gradle.kts
@@ -21,6 +21,7 @@ plugins {
 dependencies {
     runtimeOnly(project(":core:federated-catalog"))
     runtimeOnly(project(":extensions:api:federated-catalog-api"))
+    runtimeOnly(project(":extensions:store:fcc-store-sql"))
     runtimeOnly(project(":extensions:store:postgres-flyway"))
     runtimeOnly(edc.bundles.connector)
 }
